Added modified app source [mark 2]
This app source fixes some major errors and exception related issues, introduces flexible global store, has a fully operational and almost bug free  hashing programs which further increase the security of the app. Also includes fully functional app apk. If a glitch or error is found, be sure to raise an issue with the glitch tag